Make session more robust
This means that now the session will store a central object that relates rooms and users.
A person with a specific http session can be part of several rooms at the same time with different users,
something which was not possible before.
Most likely some bugs where a user was not able to vote anymore after he visited another poker room
is now fixed, since before we were not as precise in storing multiple users in the session (we were assuming a
person attends a single planning)